fix:replyCountContext 무한 루프 오류 수정중

+++ tl-spring/src/main/java/com/tl/dto/LikeDTO.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 
 public class LikeDTO {
     private String memberId;
-    private int postNo;
+    private Long postNo;
     private int liked;
     private int totalLikes;
 }
